feat: kona-providers (#596)
* feat: kona-providers

* fix: remove to_l2_block_ref utility

* fix: no_std

* fix: nits

+++ examples/trusted-sync/src/validation.rs
@@ -79,7 +79,7 @@ impl OnlineValidator {
             },
             transactions: Some(transactions),
             no_tx_pool: Some(true),
-            gas_limit: Some(header.gas_limit as u64),
+            gas_limit: Some(header.gas_limit),
         })
     }
